move specs for withAll() and withAllRecursive() to own file
